[MOVE/#206] 마이페이지 파일 위치 이동
