+++ feature/src/main/java/com/terning/feature/main/MainNavigator.kt
@@ -11,7 +11,7 @@ import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navOptions
 import com.terning.feature.calendar.calendar.navigation.navigateCalendar
 import com.terning.feature.home.home.navigation.navigateHome
-import com.terning.feature.mypage.navigation.navigateMyPage
+import com.terning.feature.mypage.mypage.navigation.navigateMyPage
 import com.terning.feature.onboarding.splash.navigation.Splash
 import com.terning.feature.search.search.navigation.navigateSearch
