+++ feature/src/main/java/com/terning/feature/main/MainTab.kt
@@ -8,7 +8,7 @@ import com.terning.core.navigation.Route
 import com.terning.feature.R
 import com.terning.feature.calendar.calendar.navigation.Calendar
 import com.terning.feature.home.home.navigation.Home
-import com.terning.feature.mypage.navigation.MyPage
+import com.terning.feature.mypage.mypage.navigation.MyPage
 import com.terning.feature.search.search.navigation.Search
 
 enum class MainTab(
