+++ feature/src/main/java/com/terning/feature/mypage/mypage/navigation/MyPageNavigation.kt
@@ -1,4 +1,4 @@
-package com.terning.feature.mypage.navigation
+package com.terning.feature.mypage.mypage.navigation
 
 import androidx.compose.animation.EnterTransition
 import androidx.compose.animation.ExitTransition
@@ -8,7 +8,7 @@ import androidx.navigation.NavHostController
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.terning.core.navigation.MainTabRoute
-import com.terning.feature.mypage.MyPageRoute
+import com.terning.feature.mypage.mypage.MyPageRoute
 import kotlinx.serialization.Serializable
 
 fun NavController.navigateMyPage(navOptions: NavOptions? = null) {
