+++ feature/src/main/java/com/terning/feature/mypage/profileedit/navigation/ProfileEditNavigation.kt
@@ -1,0 +1,4 @@
+package com.terning.feature.mypage.profileedit.navigation
+
+class ProfileEditNavigation {
+}
